ai/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==3.2.15
django-blacklist==0.6.1
django-cors-headers==3.13.0
django-filter==22.1
django-ratelimit==3.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
Faker==14.2.0
fonttools==4.37.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
mysqlclient==2.1.1
numpy==1.21.6
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
pycodestyle==2.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
